New OandaDataDownloader
