--- ToolBox/OandaDownloader/OandaRestLibrary/TradeLibrary/Communications/CandlesResponse.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OANDARestLibrary.TradeLibrary.DataTypes.Communications
-{
-    public class CandlesResponse
-    {
-	    public string instrument;
-	    public string granularity;
-        public List<Candle> candles;
-    }
-}
