--- ToolBox/OandaDownloader/OandaRestLibrary/TradeLibrary/Communications/InstrumentsResponse.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OANDARestLibrary.TradeLibrary.DataTypes.Communications
-{
-    public class InstrumentsResponse
-    {
-        public List<Instrument> instruments;
-    }
-}
