+++ ToolBox/QuantConnect.ToolBox.csproj
@@ -32,7 +32,7 @@
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
   <PropertyGroup>
-    <StartupObject>QuantConnect.ToolBox.DukascopyDownloader.Program</StartupObject>
+    <StartupObject>QuantConnect.ToolBox.OandaDownloader.Program</StartupObject>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="Ionic.Zip">
@@ -64,15 +64,13 @@
     <Compile Include="LeanBar.cs" />
     <Compile Include="LeanDataWriter.cs" />
     <Compile Include="LeanInstrument.cs" />
-    <Compile Include="OandaDownloader\ConfigSettings.cs" />
+    <Compile Include="OandaDownloader\OandaDataDownloader.cs" />
     <Compile Include="OandaDownloader\OandaRestLibrary\Credentials.cs" />
-    <Compile Include="OandaDownloader\OandaRestLibrary\Rest.cs" />
-    <Compile Include="OandaDownloader\OandaRestLibrary\TradeLibrary\Candle.cs" />
-    <Compile Include="OandaDownloader\OandaRestLibrary\TradeLibrary\Communications\CandlesResponse.cs" />
-    <Compile Include="OandaDownloader\OandaRestLibrary\TradeLibrary\Communications\InstrumentsResponse.cs" />
-    <Compile Include="OandaDownloader\OandaRestLibrary\TradeLibrary\Communications\Requests\CandlesRequest.cs" />
-    <Compile Include="OandaDownloader\OandaRestLibrary\TradeLibrary\Communications\Requests\Request.cs" />
-    <Compile Include="OandaDownloader\OandaRestLibrary\TradeLibrary\Instrument.cs" />
+    <Compile Include="OandaDownloader\OandaRestLibrary\Candle.cs" />
+    <Compile Include="OandaDownloader\OandaRestLibrary\CandlesResponse.cs" />
+    <Compile Include="OandaDownloader\OandaRestLibrary\CandlesRequest.cs" />
+    <Compile Include="OandaDownloader\OandaRestLibrary\Request.cs" />
+    <Compile Include="OandaDownloader\OandaRestLibrary\Instrument.cs" />
     <Compile Include="OandaDownloader\Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="QuantQuoteConverter\Program.cs" />
@@ -84,13 +82,10 @@
     <None Include="CoarseUniverseGenerator\config.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Include="OandaDownloader\config.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="OandaDownloader\instruments.txt">
+    <None Include="instruments_oanda.txt">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
@@ -122,6 +117,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
